+++ bumble/transport/tcp_server.py
@@ -30,6 +30,7 @@ logger = logging.getLogger(__name__)
 
 # -----------------------------------------------------------------------------
 
+
 # A pass-through function to ease mock testing.
 async def _create_server(*args, **kw_args):
     await asyncio.get_running_loop().create_server(*args, **kw_args)
