+++ rust/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Next
+
+- GATT server support
+- Tests w/ rootcanal
+- Battery service example
+- Address is now pure Rust that's convertible to its Python equivalent rather than just holding a PyObject
+
 # 0.2.0
 
 - Code-gen company ID table
