+++ rust/src/wrapper/hci.rs
@@ -14,15 +14,13 @@
 
 //! HCI
 
-use std::fmt;
 // re-export here, and internal usages of these imports should refer to this mod, not the internal
 // mod
 pub(crate) use crate::internal::hci::WithPacketType;
-pub use crate::internal::hci::{packets, Error, Packet};
+pub use crate::internal::hci::{packets, Address, Error, InvalidAddressHex, Packet};
 
 use crate::wrapper::core::{TryFromPy, TryToPy};
 use crate::wrapper::hci::packets::{AddressType, Command, ErrorCode};
-use itertools::Itertools as _;
 use pyo3::types::PyBytes;
 use pyo3::{
     exceptions::PyException, intern, types::PyModule, FromPyObject, IntoPy, PyAny, PyErr, PyObject,
@@ -72,111 +70,11 @@ impl IntoPy<PyObject> for HciCommand {
     }
 }
 
-/// A Bluetooth address
-#[derive(Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Address {
-    /// Little-endian bytes
-    le_bytes: [u8; 6],
-    address_type: AddressType,
-}
-
-impl Address {
-    /// Creates a new address with the provided little-endian bytes.
-    pub fn from_le_bytes(le_bytes: [u8; 6], address_type: AddressType) -> Self {
-        Self {
-            le_bytes,
-            address_type,
-        }
-    }
-
-    /// Creates a new address with the provided big endian hex (with or without `:` separators).
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use bumble::{wrapper::{hci::{Address, packets::AddressType}}};
-    /// let hex = "F0:F1:F2:F3:F4:F5";
-    /// assert_eq!(
-    ///     hex,
-    ///     Address::from_be_hex(hex, AddressType::PublicDeviceAddress).unwrap().as_be_hex()
-    /// );
-    /// ```
-    pub fn from_be_hex(
-        address: &str,
-        address_type: AddressType,
-    ) -> Result<Self, InvalidAddressHex> {
-        let filtered: String = address.chars().filter(|c| *c != ':').collect();
-        let mut bytes: [u8; 6] = hex::decode(filtered)
-            .map_err(|_| InvalidAddressHex { address })?
-            .try_into()
-            .map_err(|_| InvalidAddressHex { address })?;
-        bytes.reverse();
-
-        Ok(Self {
-            le_bytes: bytes,
-            address_type,
-        })
-    }
-
-    /// The type of address
-    pub fn address_type(&self) -> AddressType {
-        self.address_type
-    }
-
-    /// True if the address is static
-    pub fn is_static(&self) -> bool {
-        !self.is_public() && self.le_bytes[5] >> 6 == 3
-    }
-
-    /// True if the address type is [AddressType::PublicIdentityAddress] or
-    /// [AddressType::PublicDeviceAddress]
-    pub fn is_public(&self) -> bool {
-        matches!(
-            self.address_type,
-            AddressType::PublicDeviceAddress | AddressType::PublicIdentityAddress
-        )
-    }
-
-    /// True if the address is resolvable
-    pub fn is_resolvable(&self) -> bool {
-        matches!(
-            self.address_type,
-            AddressType::PublicIdentityAddress | AddressType::RandomIdentityAddress
-        )
-    }
-
-    /// Address bytes in _little-endian_ format
-    pub fn as_le_bytes(&self) -> [u8; 6] {
-        self.le_bytes
-    }
-
-    /// Address bytes as big-endian colon-separated hex
-    pub fn as_be_hex(&self) -> String {
-        self.le_bytes
-            .into_iter()
-            .rev()
-            .map(|byte| hex::encode_upper([byte]))
-            .join(":")
-    }
-}
-
-// show a more readable form than default Debug for a byte array
-impl fmt::Debug for Address {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "Address {{ address: {}, type: {:?} }}",
-            self.as_be_hex(),
-            self.address_type
-        )
-    }
-}
-
 impl TryToPy for Address {
     fn try_to_py<'py>(&self, py: Python<'py>) -> PyResult<&'py PyAny> {
         PyModule::import(py, intern!(py, "bumble.device"))?
             .getattr(intern!(py, "Address"))?
-            .call1((PyBytes::new(py, &self.le_bytes), self.address_type))
+            .call1((PyBytes::new(py, &self.as_le_bytes()), self.address_type()))
     }
 }
 
@@ -194,18 +92,8 @@ impl TryFromPy for Address {
 
         let address = obj.call_method0(intern!(py, "to_bytes"))?.extract()?;
 
-        Ok(Self {
-            le_bytes: address,
-            address_type,
-        })
+        Ok(Self::from_le_bytes(address, address_type))
     }
-}
-
-/// Error type for [Address::from_be_hex].
-#[derive(Debug, thiserror::Error)]
-#[error("Invalid address hex: {address}")]
-pub struct InvalidAddressHex<'a> {
-    address: &'a str,
 }
 
 /// An error meaning that the internal u64 value used to convert to [packets::Address] did not
